[fix]: PaymentHistory와 Payment 통합
