--- src/main/java/com/nodaji/payment/service/BuyService.java
@@ -1,8 +1,0 @@
-package com.nodaji.payment.service;
-
-import com.nodaji.payment.dto.request.BuyRequestDto;
-import com.nodaji.payment.dto.response.BuyResponseDto;
-
-public interface BuyService {
-    BuyResponseDto buyItems(String userId, BuyRequestDto req);
-}
