--- src/main/java/com/nodaji/payment/service/PaymentHistoryService.java
@@ -1,7 +1,0 @@
-package com.nodaji.payment.service;
-
-import org.json.simple.JSONObject;
-
-public interface PaymentHistoryService {
-    void createPaymentHistory(JSONObject jsonObject, String userId);
-}
